requirements.txt
Flask==2.0.1
gunicorn==20.1.0
opencv-python-headless==4.2.0.32
tensorflow-cpu==2.5.0

